+++ spec/features/transfers/users_can_edit_project_details_spec.rb
@@ -149,4 +149,39 @@ RSpec.feature "Users can edit transfer project details" do
     expect(page).to have_link "View the trust SharePoint folder (opens in a new tab)",
       href: "https://educationgovuk-my.sharepoint.com/incoming-trust-folder-updated-link"
   end
+
+  scenario "they can change the two requires improvement response" do
+    visit project_information_path(project)
+
+    row = find("#reasonsFor .govuk-summary-list__row:nth-child(1)")
+
+    within(row) do
+      expect(page).to have_content("No")
+      click_link "Change"
+    end
+
+    within("#two-requires-improvement") do
+      choose("Yes")
+    end
+
+    click_on "Continue"
+
+    within(row) do
+      expect(page).to have_content("Yes")
+    end
+
+    within(row) do
+      click_link "Change"
+    end
+
+    within("#two-requires-improvement") do
+      choose("No")
+    end
+
+    click_on "Continue"
+
+    within(row) do
+      expect(page).to have_content("No")
+    end
+  end
 end